+++ components/ReviewCard/reviewCard.module.scss
@@ -1,4 +1,4 @@
-@use './colors.module.scss';
+@use '../../scss/colors.module.scss';
 
 .card__descAccordion {
   z-index: 3;
